skillMvc commit

+++ JobWebApi/Controllers/JobController.cs
@@ -78,8 +78,22 @@ namespace JobWebApi.Controllers
         {
             try
             {
-                await repo.RemoveJobDetailsAsync(jobId);
-                return Ok();
+                HttpClient client = new HttpClient() { BaseAddress = new Uri("http://localhost:5117/api/JobPost/") };
+                var response = await client.DeleteAsync("Job/" +jobId);
+                HttpClient client2 = new HttpClient() { BaseAddress = new Uri("http://localhost:5210/api/JobSkill/") };
+                var response1 = await client2.DeleteAsync("Job/" + jobId);
+                HttpClient client3 = new HttpClient() { BaseAddress = new Uri("http://localhost:5005/api/Employee/") };
+                var response2 = await client3.DeleteAsync("Job/" + jobId);
+                if (response.IsSuccessStatusCode && response1.IsSuccessStatusCode && response2.IsSuccessStatusCode )
+                {
+                    await repo.RemoveJobDetailsAsync(jobId);
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest("Cannot delete the job");
+                }
+               
             }
             catch(JobException ex) 
             {
